API integration in search bar in homepage and navigation in popular service component
