+++ src/navigators/Navigator.js
@@ -42,7 +42,7 @@ import CheckoutScreen from '../screens/AuthNav/CheckOut';
 import HiringConfirmation from '../screens/AuthNav/HiringConfirmation';
 import AddExperienceNew from '../screens/AuthNav/AddExperienceNew';
 import JobListingScreen from '../screens/AuthNav/JobListing';
-import LatestProjectList from '../screens/AuthNav/LatestProjectList';
+import LatestProjectList from '../screens/NonAuthNav/LatestProjectList';
 //non stack pages
 import SignUpScreen from '../screens/NonAuthNav/SignUp';
 import SignInScreen from '../screens/NonAuthNav/SignIn';
